End of Game Jam
Played all the way through, though with extra upgrades. Game seems balanced enough, though could use a lot of fine-tuning. Final boss works when tested on level 1, hope it works on level 15 as well. End screen is rough, but I made it in like 5 min so I think it's OK. Ended up falling back on my old ways during the last week or so. Would love to see what this can be when proper OOAD is adhered to.